refactor: rename plugin for clarity
